added rating to separate widget
